Remove static shared ID cache
The cache might be reset while another thread is using it, the list will hold the spinner entities (and with it, the entire level) until you encounter another spinner breaking ball to reset it, etc.
Having a cache instead of reflection each time is good though, so I'm turning that into a local variable in the task.
Since DynData seems to be causing issues, I'm turning that into plain old reflection instead.